+++ go/cs/reservation/segment/admission/stateless/stateless.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package impl
+package stateless
 
 import (
 	"context"
